refactor(ddd): rename domain/core folder to domain/shared
