+++ lib/domain/shared/week_enum.freezed.dart
