+++ lib/domain/workout/deload_workout.dart
@@ -1,8 +1,8 @@
 import 'package:dartz/dartz.dart';
 import 'package:flutter/foundation.dart';
 
-import 'package:power_progress/domain/core/value_objects/month.dart';
-import 'package:power_progress/domain/core/value_objects/one_rm.dart';
+import 'package:power_progress/domain/shared/value_objects/month.dart';
+import 'package:power_progress/domain/shared/value_objects/one_rm.dart';
 import 'package:power_progress/domain/workout/exercise_set.dart';
 import 'package:power_progress/domain/workout/value_objects/weight.dart';
 import 'package:power_progress/domain/workout/workout.dart';
