+++ lib/domain/workout/i_workout_repository.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 
-import 'package:power_progress/domain/core/value_objects/month.dart';
-import 'package:power_progress/domain/core/week_enum.dart';
+import 'package:power_progress/domain/shared/value_objects/month.dart';
+import 'package:power_progress/domain/shared/week_enum.dart';
 import 'package:power_progress/domain/workout/workout_done.dart';
 import 'package:power_progress/domain/workout/workout_failure.dart';
