+++ lib/infrastructure/one_rm/repositories/one_rm_repository.dart
@@ -3,8 +3,8 @@ import 'package:flutter/foundation.dart';
 
 import 'package:power_progress/core/domain/errors.dart';
 import 'package:power_progress/domain/one_rm/one_rm_failure.dart';
-import 'package:power_progress/domain/core/value_objects/one_rm.dart';
-import 'package:power_progress/domain/core/value_objects/month.dart';
+import 'package:power_progress/domain/shared/value_objects/one_rm.dart';
+import 'package:power_progress/domain/shared/value_objects/month.dart';
 import 'package:power_progress/domain/one_rm/i_one_rm_repository.dart';
 import 'package:power_progress/infrastructure/one_rm/datasources/i_one_rm_datasource.dart';
 import 'package:power_progress/infrastructure/one_rm/models/one_rm_model.dart';
