+++ lib/infrastructure/workout/models/workout_done_model.dart
@@ -2,8 +2,8 @@ import 'package:flutter/foundation.dart';
 import 'package:dartz/dartz.dart';
 import 'package:hive/hive.dart';
 
-import 'package:power_progress/domain/core/value_objects/month.dart';
-import 'package:power_progress/domain/core/week_enum.dart';
+import 'package:power_progress/domain/shared/value_objects/month.dart';
+import 'package:power_progress/domain/shared/week_enum.dart';
 import 'package:power_progress/domain/workout/workout_done.dart';
 import 'package:power_progress/domain/workout/workout_failure.dart';
